allow filtering on submaps, as well as entries
you can now pass 'filter_indexes: true' to parse and discover, which
will pass the submap instances through your block for filtering as well
as your entries.

it's up to calling code to ensure the block won't break given two
different structs, but they share `loc` and `lastmod` attributes, so
that's a non-trivial ability to filter that's added.